Se actualiza la pantalla de loguin

+++ src/scripts/loguin.js
@@ -12,13 +12,55 @@ document.getElementById("loginForm").addEventListener("submit", async (e) => {
     });
 
     const data = await res.json();
+    const tokena = localStorage.getItem("tokena");
+    const token = localStorage.getItem("token");
+    console.log(token);
+    if (token || tokena) {
+      Swal.fire({
+        title: "Usuario ya logueado",
+        text: "Desea cerrar la sesion",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#3085d6",
+        cancelButtonColor: "#d33",
+        confirmButtonText: "Se, cerrar",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          localStorage.clear();
+          Swal.fire({
+            title: "Sesion finalizada!",
+            text: "Su sesion se cerro con exito",
+            icon: "success",
+          });
+        }
+      });
+      return;
+    }
+    const payloadBase64 = data.token.split(".")[1];
+    const payloadJson = atob(payloadBase64);
+    const payload = JSON.parse(payloadJson);
+
+    const rol = payload.usuario.rol;
+
+    if (rol === 1) {
+      localStorage.setItem("tokena", data.token);
+      window.location.href = "/src/Pages/gestorUsuarios.html";
+      return;
+    }
 
     if (res.ok) {
       localStorage.setItem("token", data.token);
-      console.log(data.token);
       window.location.href = "/src/Pages/gestorUsuarios.html";
     } else {
-      alert(data.mensaje || "Credenciales inválidas");
+      Toastify({
+        text: data.mensaje || "Credenciales inválidas",
+        duration: 1000,
+        position: "center",
+        style: {
+          background:
+            "linear-gradient(to right,rgb(255, 0, 51),rgb(201, 61, 61))",
+        },
+      }).showToast();
     }
   } catch (error) {
     alert("Error al conectar con el servidor");
